requirements.txt
apache-airflow==3.0.0
pandas
pyodbc
scikit-learn
requests